app-template/app/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-cloud-secret-manager==2.20.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
